+++ src/lib.rs
@@ -9,6 +9,8 @@ mod error;
 /// `Cargo.toml`
 mod error_kind;
 mod map_add_err;
+use alloc::boxed::Box;
+
 pub use error::Error;
 pub use error_kind::ErrorKind;
 pub use map_add_err::MapAddError;
@@ -49,6 +51,24 @@ macro_rules! x {
     };
 }
 
+#[allow(unused_macros)]
+macro_rules! x_box {
+    ($kind:ident $x:ty) => {
+        impl From<$x> for ErrorKind {
+            fn from(e: $x) -> Self {
+                Self::$kind(Box::new(e))
+            }
+        }
+
+        impl From<$x> for Error {
+            #[track_caller]
+            fn from(e: $x) -> Self {
+                Self::from_kind(e)
+            }
+        }
+    };
+}
+
 type X0 = ();
 unit_x!(UnitError X0);
 type X1 = &'static str;
@@ -72,7 +92,7 @@ x!(SerdeJsonError X7);
 #[cfg(feature = "ron_support")]
 type X8 = ron::error::Error;
 #[cfg(feature = "ron_support")]
-x!(RonError X8);
+x_box!(RonError X8);
 #[cfg(feature = "ctrlc_support")]
 type X9 = ctrlc::Error;
 #[cfg(feature = "ctrlc_support")]
@@ -83,16 +103,16 @@ type X11 = core::num::ParseFloatError;
 x!(ParseFloatError X11);
 type X12 = core::num::TryFromIntError;
 x!(TryFromIntError X12);
-type X13 = alloc::boxed::Box<dyn std::error::Error + Send + Sync>;
+type X13 = Box<dyn std::error::Error + Send + Sync>;
 x!(BoxedError X13);
 #[cfg(feature = "toml_support")]
 type X14 = toml::de::Error;
 #[cfg(feature = "toml_support")]
-x!(TomlDeError X14);
+x_box!(TomlDeError X14);
 #[cfg(feature = "toml_support")]
 type X15 = toml::ser::Error;
 #[cfg(feature = "toml_support")]
-x!(TomlSerError X15);
+x_box!(TomlSerError X15);
 #[cfg(feature = "serde_yaml_support")]
 type X16 = serde_yaml::Error;
 #[cfg(feature = "serde_yaml_support")]